\name{click.2-package}
\alias{click.2-package}
\docType{package}
\title{
\packageTitle{click.2}
}
\description{
\packageDescription{click.2}
}
\details{

\packageDESCRIPTION{click.2}
\packageIndices{click.2}
See the member functions for addtional details.
%%~~ An overview of how to use the package, including the most ~~
%%~~ important functions ~~
}
\author{
\packageAuthor{click.2}

Maintainer: \packageMaintainer{click.2}
}
\references{
%%~~ Literature or other references for background information ~~
}


\keyword{ package }
\keyword{ device }
\keyword{ IO }
%~~ Optionally other standard keywords, one per line, from file ~~
%~~ KEYWORDS in the R documentation directory ~~
\seealso{
\link{getGraphicsEvent}
}
